Making an effort to be more organized.
