+++ app/src/main/AndroidManifest.xml
@@ -11,18 +11,18 @@
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
         <activity
-            android:name=".MainActivity"
+            android:name=".Activities.MainActivity"
             android:label="@string/app_name">
             <intent-filter>
             <action android:name="android.intent.action.MAIN" />
             <category android:name="android.intent.category.LAUNCHER" />
         </intent-filter></activity>
         <activity
-            android:name=".SurveyActivity"
+            android:name=".Activities.SurveyActivity"
             android:label="Exit Poll">
         </activity>
         <activity
-            android:name=".StatsActivity"
+            android:name=".Activities.StatsActivity"
             android:label="Exit Poll Stats" />
     </application>
