+++ app/src/main/java/com/example/worldcom/movieexitpoller/Activities/StatsActivity.java
@@ -1,4 +1,4 @@
-package com.example.worldcom.movieexitpoller;
+package com.example.worldcom.movieexitpoller.Activities;
 
 import android.content.Intent;
 import android.graphics.Color;
@@ -7,11 +7,12 @@ import android.os.Bundle;
 
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
-import android.support.v4.content.AsyncTaskLoader;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.MenuItem;
 
+import com.example.worldcom.movieexitpoller.R;
+import com.example.worldcom.movieexitpoller.Room.ResponseRoomDatabase;
 import com.jjoe64.graphview.GraphView;
 import com.jjoe64.graphview.ValueDependentColor;
 import com.jjoe64.graphview.helper.StaticLabelsFormatter;
