+++ app/src/main/res/layout/content_main.xml
@@ -25,7 +25,7 @@ limitations under the License. -->
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryLight"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.worldcom.movieexitpoller.MainActivity"
+    tools:context="com.example.worldcom.movieexitpoller.Activities.MainActivity"
     tools:showIn="@layout/activity_main">
 
     <android.support.v7.widget.RecyclerView
